[Bug #161980640]Modify users endpoints
