+++ app/api/v1/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask, jsonify, request, Blueprint
 from flask_restful import Api
 import json
-# from velidate import *
+from .views import velidate 
 from .views.order import Home
 from .views.users import Register
 from .views.users import Login
