+++ app/api/v1/views/users.py
@@ -3,8 +3,22 @@ from flask_restful import Resource
 import datetime
 from functools import wraps
 import jwt
+import re
 
 users =  {}
+class Validation():
+	def password_verify(password,confirm_password):
+		if password == confirm_password:
+			return True
+		else:
+			return False
+
+	def valid_email(email):
+		if re.match('^[_a-z0-9-]+(\.[_a-z0-9-]+)*@[a-z0-9-]+(\.[a-z0-9-]+)*(\.[a-z]{2,4})$', email) != None:
+			return True
+		else:
+			return False
+
 
 def mustlogin(d):
 	@wraps(d)
@@ -61,7 +75,7 @@ class Register(Resource):
 				return response
 			response = jsonify({
 				'status': 'ok',
-				'message': 'success ! you can now login to continue'
+				'User': users
 			})
 			return response
